refactor(client): :recycle: refactor user profile component and fix minor bugs
