+++ client/src/components/Backend/screens/UserProfile/components/UserProfileForm/index.ts
@@ -1,0 +1,1 @@
+export { default } from "./UserProfileForm";
