+++ client/src/components/Backend/screens/UserProfile/components/index.ts
@@ -2,6 +2,7 @@ export { default as LanguageBadge } from "./LanguageBadge";
 export { default as Tag } from "./Tag";
 export { default as EditButton } from "./EditButton";
 export { default as EditAvatar } from "./EditAvatar";
+export { default as UserProfileForm } from "./UserProfileForm";
 export * from "./ModalDepartments";
 export * from "./ModalLanguage";
 export * from "./ModalResetPassword";
